Layout as a polyarc

Layout as a polyarc

as_polyarc(n = 3, space = pi/3)

Arguments

n

how many arcs of this poly_arc

space

the space between each arc, default: pi/3

Value

A two-column matrix, each row giving the coordinates of a vertex, according to the ids of the vertex ids.

See also

Examples

as_polyarc()(co_net)
#>                                                    name            X
#> 1                          s__un_f__Thermomonosporaceae  1.000000000
#> 2                              s__Flavobacterium_terrae  0.999931706
#> 3                                 s__un_g__Streptomyces  0.999726835
#> 4                                    s__un_g__Rhizobium  0.999385413
#> 5                        s__Bradyrhizobium_neotropicale  0.998907487
#> 6                                s__un_g__Amycolatopsis  0.998293124
#> 7                               s__Kribbella_catacumbae  0.997542406
#> 8   s__Amycolatopsis_keratiniphila_subsp._keratiniphila  0.996655437
#> 9                               s__Kineosporia_rhamnosa  0.995632337
#> 10                     s__Pseudoduganella_violaceinigra  0.994473246
#> 11                      s__Steroidobacter_denitrificans  0.993178323
#> 12                         s__Phenylobacterium_composti  0.991747744
#> 13                                   s__un_g__Agromyces  0.990181705
#> 14                                     s__un_g__Devosia  0.988480419
#> 15                         s__un_c__Deltaproteobacteria  0.986644120
#> 16                                    s__Dongia_mobilis  0.984673058
#> 17                             s__Pseudoxanthomonas_gei  0.982567501
#> 18                     s__Pseudoxanthomonas_wuyuanensis  0.980327738
#> 19                     s__Steroidobacter_agariperforans  0.977954075
#> 20                                   s__Devosia_insulae  0.975446836
#> 21                                s__Nocardioides_albus  0.972806362
#> 22                               s__un_g__Mesorhizobium  0.970033016
#> 23                      s__Phyllobacterium_bourgognense  0.967127176
#> 24                                   s__un_g__Niastella  0.964089238
#> 25                           s__Lysobacter_panaciterrae  0.960919618
#> 26                                 s__un_p__Chloroflexi  0.957618748
#> 27                            s__Hyphomicrobium_vulgare  0.954187079
#> 28                       s__Verrucosispora_andamanensis  0.950625081
#> 29                                       s__Bosea_eneae  0.946933239
#> 30                          s__Permianibacter_aggregans  0.943112058
#> 31                                  s__un_g__Hyphomonas  0.939162060
#> 32                        s__Sphingomonas_daechungensis  0.935083785
#> 33                              s__Caulobacter_henricii  0.930877788
#> 34                              s__un_f__Cryomorphaceae  0.926544646
#> 35                                s__un_c__Chloroflexia  0.922084949
#> 36                           s__Ohtaekwangia_kribbensis  0.917499307
#> 37                            s__Bacillus_thuringiensis  0.912788346
#> 38                                s__un_o__Myxococcales  0.907952709
#> 39                                 s__un_g__Roseiflexus  0.902993058
#> 40                             s__un_g__Solirubrobacter  0.897910070
#> 41                              s__un_g__Oscillochloris  0.892704438
#> 42                       s__Rhodanobacter_denitrificans  0.887376874
#> 43                        s__Aeromicrobium_panaciterrae  0.881928105
#> 44                          s__un_c__Betaproteobacteria  0.876358877
#> 45                              s__Sorangium_cellulosum  0.870669948
#> 46                            s__Brevundimonas_basaltis  0.864862097
#> 47                                  s__un_g__Glycomyces  0.858936117
#> 48                              s__un_g__Hyphomicrobium  0.852892817
#> 49                      s__Streptomyces_bangladeshensis  0.846733023
#> 50                              s__un_g__Bradyrhizobium  0.840457576
#> 51                           s__Paraherbaspirillum_soli  0.834067332
#> 52                               s__un_g__Mycobacterium  0.827563166
#> 53                         s__un_c__Alphaproteobacteria  0.820945965
#> 54                                   s__Acidovorax_soli  0.814216633
#> 55                               s__un_g__Paenibacillus  0.807376089
#> 56                               s__un_g__Pedomicrobium  0.800425268
#> 57                           s__un_f__Xanthobacteraceae  0.793365119
#> 58                               s__Rhizocola_hellebori  0.786196607
#> 59                               s__un_g__Lechevalieria  0.778920709
#> 60                   s__un_g__Clostridium_sensu_stricto  0.771538422
#> 61                             s__Nocardioides_jensenii  0.764050751
#> 62                             s__Lysobacter_panacisoli  0.756458721
#> 63                              s__Shinella_kummerowiae  0.748763368
#> 64                           s__Arenimonas_daejeonensis  0.740965744
#> 65                                s__Polyangium_fumosum  0.733066913
#> 66                                        s__un_g__Gp10  0.725067954
#> 67                    s__Sphingosinicella_vermicomposti  0.716969960
#> 68                             s__un_p__Verrucomicrobia  0.708774037
#> 69                          s__Herbaspirillum_aquaticum  0.700481305
#> 70                              s__Aquabacterium_parvum  0.692092896
#> 71                             s__un_g__Actinophytocola  0.683609955
#> 72                                   s__un_g__Ralstonia  0.675033642
#> 73                          s__un_o__Desulfuromonadales  0.666365128
#> 74                                         s__un_g__Gp3  0.657605597
#> 75                      s__Pseudonocardia_xinjiangensis  0.648756245
#> 76                                  s__un_g__Acidovorax  0.639818281
#> 77                                 s__un_g__Vasilyevaea  0.630792926
#> 78                                s__un_g__Conexibacter  0.621681413
#> 79                    s__Rhodopseudomonas_rhenobacensis  0.612484986
#> 80                               s__un_g__Methylotenera  0.603204901
#> 81                         s__Hyphomicrobium_zavarzinii  0.593842426
#> 82                               s__un_f__Myxococcaceae  0.584398840
#> 83                        s__Flavobacterium_saliperosum  0.574875432
#> 84                           s__Microvirga_guangxiensis  0.565273503
#> 85                       s__Mesorhizobium_silamurunense  0.555594366
#> 86                                 s__Agromyces_indicus  0.545839341
#> 87                        s__Chryseobacterium_wanjuense  0.536009761
#> 88                                     s__un_g__Cedecea  0.526106969
#> 89                 s__Cryptosporangium_minutisporangium  0.516132317
#> 90                        s__Mycobacterium_canariasense  0.506087169
#> 91                             s__un_f__Lachnospiraceae  0.495972895
#> 92                               s__Devosia_riboflavina  0.485790878
#> 93                      s__Paenibacillus_pectinilyticus  0.475542507
#> 94                                s__un_o__Chlamydiales  0.465229184
#> 95                                s__un_g__Cellulomonas  0.454852316
#> 96                               s__un_g__Luteolibacter  0.444413321
#> 97                             s__un_f__Spirochaetaceae  0.433913625
#> 98                               s__un_g__Brevibacillus  0.423354661
#> 99                            s__Arenimonas_oryziterrae  0.412737873
#> 100                          s__Nocardioides_daeguensis  0.402064710
#> 101                         s__Duganella_phyllosphaerae  0.391336630
#> 102                               s__un_g__Methylovorus  0.380555098
#> 103                          s__un_f__Flavobacteriaceae  0.369721588
#> 104                         s__Paenibacillus_prosopidis  0.358837578
#> 105                              s__un_p__Acidobacteria  0.347904555
#> 106                         s__un_o__Sphingobacteriales  0.336924013
#> 107                             s__Phytohabitans_flavus  0.325897451
#> 108    s__Streptomyces_chrysomallus_subsp._chrysomallus  0.314826376
#> 109                           s__Turicibacter_sanguinis  0.303712300
#> 110                                  s__un_f__Iamiaceae  0.292556740
#> 111         s__un_g__Subdivision3_genera_incertae_sedis  0.281361221
#> 112                       s__Actinophytocola_timorensis  0.270127271
#> 113                          s__un_f__Parachlamydiaceae  0.258856425
#> 114                    s__Methyloversatilis_universalis  0.247550223
#> 115                      s__Lechevalieria_xinjiangensis  0.236210209
#> 116                       s__un_f__Propionibacteriaceae  0.224837931
#> 117                    s__Catelliglobosispora_koreensis  0.213434943
#> 118                         s__un_f__Bacteriovoracaceae  0.202002803
#> 119                        s__Nocardioides_iriomotensis  0.190543072
#> 120                           s__Brevundimonas_faecalis  0.179057315
#> 121                            s__Bacteriovorax_stolpii  0.167547101
#> 122                        s__Paenibacillus_brasilensis  0.156014002
#> 123                                s__Caenimonas_terrae  0.144459593
#> 124                        s__Bradyrhizobium_daqingense  0.132885454
#> 125                            s__un_f__Sinobacteraceae  0.121293163
#> 126                             s__Cystobacter_gracilis  0.109684306
#> 127                          s__Bacillus_kokeshiiformis  0.098060467
#> 128                           s__un_g__Pseudorhodoferax  0.086423234
#> 129                               s__un_g__Enterobacter  0.074774197
#> 130                             s__Agromyces_subbeticus  0.063114947
#> 131                         s__Blastocatella_fastidiosa  0.051447076
#> 132                               s__un_g__Chitinophaga  0.039772178
#> 133                         s__Arthrobacter_globiformis  0.028091848
#> 134                                        s__un_g__Gp5  0.016407681
#> 135                           s__un_g__Stenotrophomonas  0.004721272
#> 136                           s__un_o__Rhodospirillales -0.006965781
#> 137                    s__Aciditerrimonas_ferrireducens -0.018651883
#> 138                           s__Aridibacter_famidurans -0.030335437
#> 139                 s__Altererythrobacter_xinjiangensis -0.042014848
#> 140                           s__un_f__Chitinophagaceae -0.053688520
#> 141                                        s__un_g__Gp7 -0.065354859
#> 142                       s__Nocardioides_marinquilinus -0.077012271
#> 143                           s__Blastococcus_jejuensis -0.088659164
#> 144                         s__Micromonospora_siamensis -0.100293948
#> 145                               s__Clostridium_roseum -0.111915033
#> 146                         s__un_f__Gracilibacteraceae -0.123520831
#> 147                           s__Flavobacterium_hauense -0.135109759
#> 148                          s__Oligoflexus_tunisiensis -0.146680231
#> 149                      s__Parachlamydia_acanthamoebae -0.158230670
#> 150                             s__Chryseolinea_serpens -0.169759496
#> 151                         s__Nitrosospira_multiformis -0.181265134
#> 152                               s__Pelomonas_puraquae -0.506687837
#> 153                                s__un_g__Rhizobacter -0.516729267
#> 154                           s__Lentzea_flaviverrucosa -0.526700112
#> 155                         s__un_g__Noviherbaspirillum -0.536599010
#> 156                           s__Streptomyces_ederensis -0.546424609
#> 157                   s__Actinophytocola_burenkhanensis -0.556175566
#> 158                                s__un_o__Rhizobiales -0.565850550
#> 159                        s__Kinneretia_asaccharophila -0.575448240
#> 160                           s__un_f__Oxalobacteraceae -0.584967323
#> 161                          s__Piscinibacter_aquaticus -0.594406500
#> 162                                s__Sphingopyxis_soli -0.603764481
#> 163                         s__un_f__Micromonosporaceae -0.613039989
#> 164                             s__un_g__Flavobacterium -0.622231755
#> 165                          s__un_f__Bradyrhizobiaceae -0.631338525
#> 166          s__Pseudomonas_chlororaphis_subsp._piscium -0.640359054
#> 167                                 s__un_g__Variovorax -0.649292111
#> 168                           s__un_f__Xanthomonadaceae -0.658136475
#> 169                                   s__un_g__Massilia -0.666890937
#> 170                        s__un_c__Gammaproteobacteria -0.675554302
#> 171                         s__un_f__Pseudonocardiaceae -0.684125388
#> 172                       s__Actinoplanes_xinjiangensis -0.692603021
#> 173                             s__Devosia_chinhatensis -0.700986046
#> 174                              s__un_g__Aquabacterium -0.709273316
#> 175                       s__Aeromicrobium_ginsengisoli -0.717463700
#> 176                           s__Massilia_namucuonensis -0.725556079
#> 177                         s__un_g__Altererythrobacter -0.733549347
#> 178                         s__Paenibacillus_panacisoli -0.741442412
#> 179                                  s__Reyranella_soli -0.749234197
#> 180                           s__un_f__Methylophilaceae -0.756923637
#> 181                                s__Sphingomonas_soli -0.764509682
#> 182                             s__un_g__Hydrogenophaga -0.771991294
#> 183                            s__un_f__Kineosporiaceae -0.779367453
#> 184                         s__Saccharothrix_longispora -0.786637151
#> 185                                   s__un_g__Nocardia -0.793799395
#> 186                   s__Promicromonospora_xylanilytica -0.800853206
#> 187                                  s__Gaiella_occulta -0.807797621
#> 188                        s__Streptomyces_acidiscabies -0.814631691
#> 189                             s__Peredibacter_starrii -0.821354483
#> 190                               s__un_g__Burkholderia -0.827965078
#> 191                            s__Devosia_yakushimensis -0.834462573
#> 192                            s__Lentzea_kentuckyensis -0.840846081
#> 193                        s__un_f__Sphingobacteriaceae -0.847114731
#> 194                        s__un_o__Solirubrobacterales -0.853267664
#> 195                              s__un_g__Aeromicrobium -0.859304042
#> 196                               s__un_f__Rhodobiaceae -0.865223039
#> 197                         s__Asticcacaulis_solisilvae -0.871023847
#> 198                            s__Aquabacterium_commune -0.876705674
#> 199                    s__Phyllobacterium_myrsinacearum -0.882267743
#> 200                            s__Rhizobium_subbaraonis -0.887709295
#> 201                             s__Inhella_inkyongensis -0.893029586
#> 202                           s__un_o__Acidimicrobiales -0.898227889
#> 203                           s__Mesorhizobium_gobiense -0.903303495
#> 204                             s__Conexibacter_arvalis -0.908255711
#> 205                     s__Hyphomicrobium_sulfonivorans -0.913083859
#> 206                                 s__Devosia_pacifica -0.917787280
#> 207                             s__Rhizobium_laguerreae -0.922365332
#> 208                          s__un_f__Sphingomonadaceae -0.926817389
#> 209                          s__Cellvibrio_ostraviensis -0.931142843
#> 210                                  s__un_g__Geobacter -0.935341104
#> 211                          s__un_f__Planctomycetaceae -0.939411597
#> 212                              s__Planifilum_composti -0.943353768
#> 213                              s__un_g__Ilumatobacter -0.947167076
#> 214                          s__Rubrobacter_bracarensis -0.950851002
#> 215                                  s__Variovorax_soli -0.954405043
#> 216                              s__un_f__Bacillaceae_1 -0.957828712
#> 217                          s__un_f__Rhodospirillaceae -0.961121542
#> 218                         s__Blastococcus_saxobsidens -0.964283083
#> 219                            s__Cellulomonas_humilata -0.967312904
#> 220       s__un_g__Spartobacteria_genera_incertae_sedis -0.970210590
#> 221                        s__Ferrovibrio_denitrificans -0.972975746
#> 222                                 s__Pseudomonas_soli -0.975607994
#> 223                         s__Rhizobium_rosettiformans -0.978106974
#> 224                             s__Paenibacillus_terrae -0.980472345
#> 225                                        s__un_g__Gp6 -0.982703785
#> 226                      s__Solirubrobacter_taibaiensis -0.984800987
#> 227                            s__Shinella_zoogloeoides -0.986763666
#> 228                               s__un_g__Blastococcus -0.988591554
#> 229                                 s__un_g__Haliangium -0.990284400
#> 230                                 s__Rhizobium_vignae -0.991841974
#> 231                                    s__un_g__Pantoea -0.993264063
#> 232                       s__Brevibacillus_ginsengisoli -0.994550473
#> 233                                 s__un_g__Dokdonella -0.995701027
#> 234                            s__un_f__Nocardioidaceae -0.996715569
#> 235                             s__un_f__Rhodocyclaceae -0.997593961
#> 236                                 s__un_o__Bacillales -0.998336081
#> 237                        s__Chryseobacterium_defluvii -0.998941829
#> 238                                 s__un_g__Microvirga -0.999411122
#> 239                                s__un_g__Marmoricola -0.999743896
#> 240                             s__un_p__Actinobacteria -0.999940106
#> 241                              s__un_g__Fictibacillus -0.999999724
#> 242                              s__Burkholderia_heleia -0.999922743
#> 243                             s__Cellulomonas_hominis -0.999709173
#> 244                                s__un_g__Cystobacter -0.999359042
#> 245                              s__Streptomyces_bullii -0.998872400
#> 246                               s__Bauldia_consociata -0.998249313
#> 247                                s__un_g__Caulobacter -0.997489865
#> 248                              s__un_p__Bacteroidetes -0.996594160
#> 249                        s__Brevundimonas_aveniformis -0.995562321
#> 250                              s__Marmoricola_scoriae -0.994394489
#> 251                           s__Rudaea_cellulosilytica -0.993090823
#> 252                              s__Nannocystis_exedens -0.991651501
#> 253                               s__Haliangium_tepidum -0.990076721
#> 254                   s__Altererythrobacter_troitsensis -0.988366696
#> 255                         s__Yonghaparkia_alkaliphila -0.986521660
#> 256                   s__Ammoniibacillus_agariperforans -0.984541867
#> 257                              s__Nonomuraea_salmonea -0.982427585
#> 258                        s__Rhizomicrobium_electricum -0.980179104
#> 259                        s__Tumebacillus_ginsengisoli -0.977796731
#> 260                   s__Uliginosibacterium_gangwonense -0.975280792
#> 261                           s__un_g__Plantactinospora -0.972631630
#> 262                              s__Cupriavidus_necator -0.969849606
#> 263                                   s__Bacillus_horti -0.966935102
#> 264                                 s__un_g__Planifilum -0.963888515
#> 265                                       s__un_g__Gp11 -0.960710261
#> 266                          s__un_f__Microbacteriaceae -0.957400774
#> 267                          s__Inquilinus_ginsengisoli -0.953960507
#> 268                               s__Byssovorax_cruenta -0.950389930
#> 269                       s__Sporacetigenium_mesophilum -0.946689529
#> 270                             s__Xenophilus_aerolatus -0.942859812
#> 271                               s__Bacillus_idriensis -0.938901300
#> 272                           s__Nocardioides_terrigena -0.934814535
#> 273                                s__Bacillus_invictae -0.930600074
#> 274                           s__un_g__Geodermatophilus -0.926258494
#> 275                               s__Arenimonas_metalli -0.921790387
#> 276                              s__un_g__Brevundimonas -0.917196364
#> 277                             s__un_f__Nannocystaceae -0.912477053
#> 278                                 s__un_g__Terrimonas -0.907633098
#> 279                             s__un_p__Planctomycetes -0.902665160
#> 280                                s__un_g__Sporichthya -0.897573919
#> 281                 s__Mangrovibacterium_diazotrophicum -0.892360070
#> 282                       s__Sphingorhabdus_planktonica -0.887024325
#> 283                          s__Nonomuraea_muscovyensis -0.881567412
#> 284                              s__un_g__Terribacillus -0.875990078
#> 285                       s__Phreatobacter_oligotrophus -0.870293084
#> 286                            s__Marmoricola_aequoreus -0.864477209
#> 287                        s__un_f__Verrucomicrobiaceae -0.858543246
#> 288                                    s__un_g__Bauldia -0.852492007
#> 289                           s__Rubrivivax_gelatinosus -0.846324317
#> 290                           s__Arenimonas_donghaensis -0.840041021
#> 291                      s__Paenibacillus_barcinonensis -0.833642975
#> 292                                s__Lysobacter_oryzae -0.827131053
#> 293                               s__un_g__Sphingopyxis -0.820506146
#> 294                             s__un_f__Leptospiraceae -0.813769159
#> 295                      s__Clostridium_straminisolvens -0.806921010
#> 296                                 s__un_g__Romboutsia -0.799962637
#> 297                          s__Reyranella_graminifolii -0.792894989
#> 298                             s__un_g__Clostridium_XI -0.785719033
#> 299                               s__un_c__Subdivision3 -0.778435747
#> 300                   s__Thioalkalivibrio_sulfidiphilus -0.771046127
#> 301                            s__un_f__Sneathiellaceae -0.763551183
#> 302                            s__Rhizobacter_bergeniae -0.496644824
#> 303                            s__un_o__Burkholderiales -0.486466796
#> 304                               s__un_g__Actinoplanes -0.476222317
#> 305                             s__un_f__Comamonadaceae -0.465912787
#> 306                           s__Actinocorallia_herbida -0.455539612
#> 307                             s__un_p__Proteobacteria -0.445104212
#> 308                               s__un_g__Sphingomonas -0.434608010
#> 309             s__un_f__Burkholderiales_incertae_sedis -0.424052441
#> 310                                   s__un_g__Bacillus -0.413438947
#> 311                                    s__un_g__Lentzea -0.402768977
#> 312                              s__un_g__Herpetosiphon -0.392043989
#> 313                                    s__un_g__Ensifer -0.381265448
#> 314                               s__un_g__Nocardioides -0.370434826
#> 315                     s__Noviherbaspirillum_suwonense -0.359553603
#> 316                  s__Kibdelosporangium_phytohabitans -0.348623266
#> 317                                s__un_g__Pseudomonas -0.337645306
#> 318                            s__un_o__Actinomycetales -0.326621225
#> 319                        s__Sandaracinus_amylolyticus -0.315552527
#> 320                            s__Lysobacter_brunescens -0.304440724
#> 321                                s__Massilia_aerilata -0.293287335
#> 322                             s__un_g__Pseudonocardia -0.282093884
#> 323                               s__un_o__Cytophagales -0.270861898
#> 324                       s__un_f__Streptosporangiaceae -0.259592913
#> 325                        s__Variovorax_boronicumulans -0.248288468
#> 326                                  s__Rhizobium_alvei -0.236950106
#> 327                                      s__un_g__Bosea -0.225579377
#> 328                             s__un_g__Steroidobacter -0.214177834
#> 329                             s__un_c__Actinobacteria -0.202747035
#> 330                         s__Nocardioides_agariphilus -0.191288540
#> 331                             s__Lentzea_jiangxiensis -0.179803916
#> 332                          s__un_g__Dactylosporangium -0.168294730
#> 333                                s__Lysobacter_terrae -0.156762555
#> 334                  s__Cellvibrio_mixtus_subsp._mixtus -0.145208967
#> 335                               s__un_g__Ohtaekwangia -0.133635543
#> 336                           s__un_g__Phenylobacterium -0.122043865
#> 337                                 s__un_g__Lysobacter -0.110435516
#> 338                            s__Polymorphospora_rubra -0.098812081
#> 339                            s__un_f__Chloroflexaceae -0.087175148
#> 340                               s__un_g__Arthrobacter -0.075526307
#> 341                                    s__Bosea_lathyri -0.063867150
#> 342                             s__Iamia_majanohamensis -0.052199268
#> 343                           s__Ilumatobacter_fluminis -0.040524256
#> 344                          s__un_g__Promicromonospora -0.028843708
#> 345                          s__un_f__Hyphomicrobiaceae -0.017159220
#> 346                           s__Lysobacter_enzymogenes -0.005472389
#> 347                             s__un_g__Micromonospora  0.006215191
#> 348                           s__Caulobacter_fusiformis  0.017901921
#> 349                         s__Pedomicrobium_manganicum  0.029586206
#> 350                                 s__un_g__Roseateles  0.041266449
#> 351                   s__Methyloceanibacter_caenitepidi  0.052941056
#> 352                                s__Turneriella_parva  0.064608430
#> 353                              s__un_f__Polyangiaceae  0.076266980
#> 354                               s__un_f__Rhizobiaceae  0.087915111
#> 355                                 s__un_g__Arenimonas  0.099551233
#> 356                               s__Bacillus_abyssalis  0.111173756
#> 357                             s__un_o__Chloroflexales  0.122781093
#> 358                                s__un_g__Sphingobium  0.134371659
#> 359                                s__Duganella_radicis  0.145943869
#> 360                         s__un_f__Phyllobacteriaceae  0.157496143
#> 361                                s__un_g__Ramlibacter  0.169026904
#> 362                              s__un_g__Achromobacter  0.180534575
#> 363           s__Couchioplanes_caeruleus_subsp._azureus  0.192017585
#> 364                         s__Roseiflexus_castenholzii  0.203474366
#> 365                            s__un_g__Intrasporangium  0.214903353
#> 366                         s__un_f__Paenibacillaceae_2  0.226302984
#> 367                              s__un_f__Chlamydiaceae  0.237671701
#> 368                         s__un_f__Paenibacillaceae_1  0.249007953
#> 369                      s__Pseudonocardia_adelaidensis  0.260310191
#> 370                         s__un_f__Enterobacteriaceae  0.271576871
#> 371                                s__Bacillus_infantis  0.282806453
#> 372                             s__Nocardioides_dilutus  0.293997404
#> 373                         s__Microbacterium_aerolatum  0.305148195
#> 374                            s__un_o__Xanthomonadales  0.316257302
#> 375                     s__Brevundimonas_kwangchunensis  0.327323210
#> 376                              s__Umezawaea_tangerina  0.338344405
#> 377                           s__un_o__Sphingomonadales  0.349319382
#> 378                          s__Terribacillus_goriensis  0.360246643
#> 379                              s__un_g__Saccharothrix  0.371124694
#> 380                        s__Cellulosimicrobium_funkei  0.381952050
#> 381                             s__un_f__Planococcaceae  0.392727231
#> 382                          s__Nocardioides_panacisoli  0.403448765
#> 383                           s__Blastopirellula_cremea  0.414115189
#> 384                                 s__un_c__Clostridia  0.424725045
#> 385                               s__un_g__Bdellovibrio  0.435276884
#> 386                  s__Paenisporosarcina_macmurdoensis  0.445769264
#> 387                           s__Paenibacillus_urinalis  0.456200752
#> 388                        s__Pseudonocardia_seranimata  0.466569923
#> 389                                       s__un_g__Gp17  0.476875362
#> 390                                s__Rhizobium_nepotum  0.487115659
#> 391                           s__un_f__Flammeovirgaceae  0.497289416
#> 392                         s__Flavobacterium_anhuiense  0.507395244
#> 393                              s__Labilithrix_luteola  0.517431762
#> 394                      s__Pseudonocardia_kunmingensis  0.527397599
#> 395                             s__Mycobacterium_vaccae  0.537291394
#> 396                                 s__un_g__Cellvibrio  0.547111795
#> 397                               s__Bacillus_nealsonii  0.556857461
#> 398                           s__un_g__Chryseobacterium  0.566527060
#> 399                          s__un_c__Acidobacteria_Gp4  0.576119272
#> 400                           s__Cryobacterium_arcticum  0.585632787
#> 401                         s__un_f__Intrasporangiaceae  0.595066304
#> 402                            s__un_g__Blastopirellula  0.604418535
#> 403                            s__un_g__Novosphingobium  0.613688203
#> 404                        s__Phytohabitans_houttuyneae  0.622874042
#> 405                          s__un_g__Paenisporosarcina  0.631974796
#> 406                         s__Ilumatobacter_nonamiense  0.640989223
#> 407                          s__Nocardioides_mesophilus  0.649916091
#> 408                    s__Microbacterium_arthrosphaerae  0.658754181
#> 409                              s__Nonomuraea_maritima  0.667502286
#> 410                           s__un_f__Cystobacteraceae  0.676159209
#> 411                    s__Roseimicrobium_gellanilyticum  0.684723770
#> 412                          s__Nocardioides_caricicola  0.693194798
#> 413                               s__Archangium_gephyra  0.701571136
#> 414                         s__Streptomyces_seranimatus  0.709851639
#> 415                               s__Bacillus_litoralis  0.718035177
#> 416                                   s__Ensifer_saheli  0.726120632
#> 417                      s__Flavobacterium_gyeonganense  0.734106899
#> 418                                   s__un_g__Cohnella  0.741992888
#> 419                         s__Azohydromonas_australica  0.749777520
#> 420                          s__Pseudonocardia_parietis  0.757459734
#> 421                              s__un_f__Bacillaceae_2  0.765038478
#> 422                            s__Solirubrobacter_pauli  0.772512719
#> 423                                s__Microvirga_lupini  0.779881435
#> 424                             s__un_f__Caldilineaceae  0.787143619
#> 425                                       s__un_g__Gp25  0.794298280
#> 426                                   s__un_g__Shinella  0.801344441
#> 427                              s__un_g__Angustibacter  0.808281138
#> 428                           s__Lysobacter_dokdonensis  0.815107424
#> 429                             s__Rhizobium_rhizoryzae  0.821822367
#> 430                                s__un_g__Cupriavidus  0.828425049
#> 431                         s__Pseudoxanthomonas_indica  0.834914569
#> 432                             s__Nocardia_salmonicida  0.841290040
#> 433                         s__un_f__Erythrobacteraceae  0.847550591
#> 434                                s__un_g__Amaricoccus  0.853695367
#> 435                            s__un_g__Gracilibacillus  0.859723528
#> 436                            s__un_g__Planosporangium  0.865634252
#> 437                              s__Microvirga_aerilata  0.871426731
#> 438                             s__un_g__Verrucosispora  0.877100173
#> 439                              s__Terrimonas_aquatica  0.882653804
#> 440                                s__un_g__Phycicoccus  0.888086864
#> 441                               s__Starkeya_koreensis  0.893398613
#> 442                              s__Nonomuraea_kuesteri  0.898588323
#> 443                              s__un_g__Methylophilus  0.903655287
#> 444                            s__Schlegelella_aquatica  0.908598812
#> 445                          s__Marmoricola_bigeumensis  0.913418223
#> 446                              s__Povalibacter_uvarum  0.918112861
#> 447                             s__Pelagibius_litoralis  0.922682085
#> 448                          s__Aggregicoccus_edonensis  0.927125272
#> 449                   s__Flavobacterium_urocaniciphilum  0.931441813
#> 450                         s__Lacibacter_daechungensis  0.935631120
#> 451                        s__Phenylobacterium_koreense  0.939692621
#>                 Y
#> 1    0.0000000000
#> 2    0.0116868611
#> 3    0.0233721259
#> 4    0.0350541984
#> 5    0.0467314829
#> 6    0.0584023845
#> 7    0.0700653090
#> 8    0.0817186635
#> 9    0.0933608563
#> 10   0.1049902972
#> 11   0.1166053977
#> 12   0.1282045714
#> 13   0.1397862339
#> 14   0.1513488034
#> 15   0.1628907006
#> 16   0.1744103490
#> 17   0.1859061751
#> 18   0.1973766088
#> 19   0.2088200833
#> 20   0.2202350356
#> 21   0.2316199066
#> 22   0.2429731413
#> 23   0.2542931889
#> 24   0.2655785032
#> 25   0.2768275429
#> 26   0.2880387714
#> 27   0.2992106574
#> 28   0.3103416751
#> 29   0.3214303040
#> 30   0.3324750295
#> 31   0.3434743432
#> 32   0.3544267426
#> 33   0.3653307318
#> 34   0.3761848214
#> 35   0.3869875289
#> 36   0.3977373788
#> 37   0.4084329028
#> 38   0.4190726400
#> 39   0.4296551372
#> 40   0.4401789489
#> 41   0.4506426377
#> 42   0.4610447744
#> 43   0.4713839382
#> 44   0.4816587169
#> 45   0.4918677071
#> 46   0.5020095144
#> 47   0.5120827535
#> 48   0.5220860485
#> 49   0.5320180332
#> 50   0.5418773510
#> 51   0.5516626551
#> 52   0.5613726091
#> 53   0.5710058866
#> 54   0.5805611720
#> 55   0.5900371600
#> 56   0.5994325564
#> 57   0.6087460779
#> 58   0.6179764523
#> 59   0.6271224189
#> 60   0.6361827286
#> 61   0.6451561436
#> 62   0.6540414385
#> 63   0.6628373996
#> 64   0.6715428255
#> 65   0.6801565271
#> 66   0.6886773278
#> 67   0.6971040640
#> 68   0.7054355845
#> 69   0.7136707514
#> 70   0.7218084399
#> 71   0.7298475385
#> 72   0.7377869491
#> 73   0.7456255873
#> 74   0.7533623824
#> 75   0.7609962778
#> 76   0.7685262307
#> 77   0.7759512126
#> 78   0.7832702093
#> 79   0.7904822212
#> 80   0.7975862632
#> 81   0.8045813650
#> 82   0.8114665711
#> 83   0.8182409411
#> 84   0.8249035497
#> 85   0.8314534869
#> 86   0.8378898580
#> 87   0.8442117840
#> 88   0.8504184013
#> 89   0.8565088621
#> 90   0.8624823347
#> 91   0.8683380030
#> 92   0.8740750673
#> 93   0.8796927440
#> 94   0.8851902658
#> 95   0.8905668817
#> 96   0.8958218574
#> 97   0.9009544752
#> 98   0.9059640339
#> 99   0.9108498494
#> 100  0.9156112543
#> 101  0.9202475983
#> 102  0.9247582480
#> 103  0.9291425874
#> 104  0.9334000177
#> 105  0.9375299572
#> 106  0.9415318421
#> 107  0.9454051255
#> 108  0.9491492785
#> 109  0.9527637898
#> 110  0.9562481655
#> 111  0.9596019297
#> 112  0.9628246245
#> 113  0.9659158095
#> 114  0.9688750626
#> 115  0.9717019796
#> 116  0.9743961744
#> 117  0.9769572790
#> 118  0.9793849435
#> 119  0.9816788364
#> 120  0.9838386443
#> 121  0.9858640723
#> 122  0.9877548437
#> 123  0.9895107002
#> 124  0.9911314021
#> 125  0.9926167279
#> 126  0.9939664748
#> 127  0.9951804584
#> 128  0.9962585129
#> 129  0.9972004911
#> 130  0.9980062642
#> 131  0.9986757223
#> 132  0.9992087739
#> 133  0.9996053462
#> 134  0.9998653849
#> 135  0.9999888547
#> 136  0.9999757387
#> 137  0.9998260385
#> 138  0.9995397747
#> 139  0.9991169864
#> 140  0.9985577313
#> 141  0.9978620859
#> 142  0.9970301450
#> 143  0.9960620224
#> 144  0.9949578504
#> 145  0.9937177796
#> 146  0.9923419795
#> 147  0.9908306380
#> 148  0.9891839615
#> 149  0.9874021750
#> 150  0.9854855218
#> 151  0.9834342637
#> 152  0.8621295934
#> 153  0.8561488566
#> 154  0.8500511702
#> 155  0.8438373673
#> 156  0.8375082966
#> 157  0.8310648227
#> 158  0.8245078257
#> 159  0.8178382013
#> 160  0.8110568606
#> 161  0.8041647299
#> 162  0.7971627507
#> 163  0.7900518795
#> 164  0.7828330875
#> 165  0.7755073609
#> 166  0.7680757004
#> 167  0.7605391211
#> 168  0.7528986525
#> 169  0.7451553383
#> 170  0.7373102362
#> 171  0.7293644179
#> 172  0.7213189688
#> 173  0.7131749878
#> 174  0.7049335875
#> 175  0.6965958935
#> 176  0.6881630449
#> 177  0.6796361936
#> 178  0.6710165043
#> 179  0.6623051544
#> 180  0.6535033340
#> 181  0.6446122453
#> 182  0.6356331029
#> 183  0.6265671333
#> 184  0.6174155749
#> 185  0.6081796779
#> 186  0.5988607038
#> 187  0.5894599256
#> 188  0.5799786274
#> 189  0.5704181045
#> 190  0.5607796627
#> 191  0.5510646187
#> 192  0.5412742995
#> 193  0.5314100425
#> 194  0.5214731951
#> 195  0.5114651147
#> 196  0.5013871685
#> 197  0.4912407330
#> 198  0.4810271942
#> 199  0.4707479474
#> 200  0.4604043966
#> 201  0.4499979548
#> 202  0.4395300435
#> 203  0.4290020925
#> 204  0.4184155401
#> 205  0.4077718324
#> 206  0.3970724232
#> 207  0.3863187741
#> 208  0.3755123540
#> 209  0.3646546391
#> 210  0.3537471126
#> 211  0.3427912643
#> 212  0.3317885910
#> 213  0.3207405955
#> 214  0.3096487869
#> 215  0.2985146805
#> 216  0.2873397972
#> 217  0.2761256633
#> 218  0.2648738108
#> 219  0.2535857767
#> 220  0.2422631029
#> 221  0.2309073360
#> 222  0.2195200274
#> 223  0.2081027323
#> 224  0.1966570106
#> 225  0.1851844256
#> 226  0.1736865445
#> 227  0.1621649379
#> 228  0.1506211796
#> 229  0.1390568466
#> 230  0.1274735184
#> 231  0.1158727775
#> 232  0.1042562083
#> 233  0.0926253979
#> 234  0.0809819348
#> 235  0.0693274096
#> 236  0.0576634144
#> 237  0.0459915423
#> 238  0.0343133879
#> 239  0.0226305462
#> 240  0.0109446132
#> 241 -0.0007428148
#> 242 -0.0124301413
#> 243 -0.0241157699
#> 244 -0.0357981043
#> 245 -0.0474755487
#> 246 -0.0591465080
#> 247 -0.0708093878
#> 248 -0.0824625952
#> 249 -0.0941045382
#> 250 -0.1057336265
#> 251 -0.1173482717
#> 252 -0.1289468872
#> 253 -0.1405278885
#> 254 -0.1520896939
#> 255 -0.1636307238
#> 256 -0.1751494019
#> 257 -0.1866441547
#> 258 -0.1981134119
#> 259 -0.2095556070
#> 260 -0.2209691768
#> 261 -0.2323525624
#> 262 -0.2437042087
#> 263 -0.2550225652
#> 264 -0.2663060856
#> 265 -0.2775532288
#> 266 -0.2887624584
#> 267 -0.2999322431
#> 268 -0.3110610572
#> 269 -0.3221473806
#> 270 -0.3331896987
#> 271 -0.3441865033
#> 272 -0.3551362922
#> 273 -0.3660375697
#> 274 -0.3768888466
#> 275 -0.3876886406
#> 276 -0.3984354766
#> 277 -0.4091278864
#> 278 -0.4197644096
#> 279 -0.4303435932
#> 280 -0.4408639920
#> 281 -0.4513241690
#> 282 -0.4617226954
#> 283 -0.4720581506
#> 284 -0.4823291229
#> 285 -0.4925342093
#> 286 -0.5026720158
#> 287 -0.5127411575
#> 288 -0.5227402590
#> 289 -0.5326679544
#> 290 -0.5425228876
#> 291 -0.5523037125
#> 292 -0.5620090929
#> 293 -0.5716377032
#> 294 -0.5811882280
#> 295 -0.5906593628
#> 296 -0.6000498138
#> 297 -0.6093582983
#> 298 -0.6185835447
#> 299 -0.6277242929
#> 300 -0.6367792942
#> 301 -0.6457473118
#> 302 -0.8679538689
#> 303 -0.8736990649
#> 304 -0.8793249140
#> 305 -0.8848306478
#> 306 -0.8902155140
#> 307 -0.8954787773
#> 308 -0.9006197186
#> 309 -0.9056376357
#> 310 -0.9105318431
#> 311 -0.9153016723
#> 312 -0.9199464717
#> 313 -0.9244656068
#> 314 -0.9288584604
#> 315 -0.9331244323
#> 316 -0.9372629399
#> 317 -0.9412734179
#> 318 -0.9451553183
#> 319 -0.9489081109
#> 320 -0.9525312832
#> 321 -0.9560243401
#> 322 -0.9593868045
#> 323 -0.9626182172
#> 324 -0.9657181367
#> 325 -0.9686861395
#> 326 -0.9715218203
#> 327 -0.9742247916
#> 328 -0.9767946843
#> 329 -0.9792311473
#> 330 -0.9815338478
#> 331 -0.9837024712
#> 332 -0.9857367213
#> 333 -0.9876363203
#> 334 -0.9894010086
#> 335 -0.9910305452
#> 336 -0.9925247075
#> 337 -0.9938832914
#> 338 -0.9951061113
#> 339 -0.9961930002
#> 340 -0.9971438095
#> 341 -0.9979584095
#> 342 -0.9986366889
#> 343 -0.9991785550
#> 344 -0.9995839337
#> 345 -0.9998527697
#> 346 -0.9999850264
#> 347 -0.9999806855
#> 348 -0.9998397478
#> 349 -0.9995622324
#> 350 -0.9991481773
#> 351 -0.9985976390
#> 352 -0.9979106928
#> 353 -0.9970874324
#> 354 -0.9961279703
#> 355 -0.9950324377
#> 356 -0.9938009841
#> 357 -0.9924337777
#> 358 -0.9909310053
#> 359 -0.9892928723
#> 360 -0.9875196023
#> 361 -0.9856114376
#> 362 -0.9835686388
#> 363 -0.9813914850
#> 364 -0.9790802736
#> 365 -0.9766353204
#> 366 -0.9740569591
#> 367 -0.9713455422
#> 368 -0.9685014399
#> 369 -0.9655250408
#> 370 -0.9624167514
#> 371 -0.9591769964
#> 372 -0.9558062182
#> 373 -0.9523048773
#> 374 -0.9486734520
#> 375 -0.9449124384
#> 376 -0.9410223502
#> 377 -0.9370037188
#> 378 -0.9328570932
#> 379 -0.9285830397
#> 380 -0.9241821421
#> 381 -0.9196550018
#> 382 -0.9150022370
#> 383 -0.9102244833
#> 384 -0.9053223934
#> 385 -0.9002966369
#> 386 -0.8951479003
#> 387 -0.8898768869
#> 388 -0.8844843167
#> 389 -0.8789709264
#> 390 -0.8733374691
#> 391 -0.8675847143
#> 392 -0.8617134479
#> 393 -0.8557244717
#> 394 -0.8496186041
#> 395 -0.8433966789
#> 396 -0.8370595462
#> 397 -0.8306080715
#> 398 -0.8240431361
#> 399 -0.8173656368
#> 400 -0.8105764858
#> 401 -0.8036766104
#> 402 -0.7966669532
#> 403 -0.7895484716
#> 404 -0.7823221381
#> 405 -0.7749889397
#> 406 -0.7675498782
#> 407 -0.7600059698
#> 408 -0.7523582449
#> 409 -0.7446077483
#> 410 -0.7367555385
#> 411 -0.7288026883
#> 412 -0.7207502841
#> 413 -0.7125994256
#> 414 -0.7043512264
#> 415 -0.6960068132
#> 416 -0.6875673258
#> 417 -0.6790339169
#> 418 -0.6704077524
#> 419 -0.6616900105
#> 420 -0.6528818821
#> 421 -0.6439845703
#> 422 -0.6349992905
#> 423 -0.6259272701
#> 424 -0.6167697483
#> 425 -0.6075279761
#> 426 -0.5982032159
#> 427 -0.5887967413
#> 428 -0.5793098374
#> 429 -0.5697438001
#> 430 -0.5600999360
#> 431 -0.5503795626
#> 432 -0.5405840075
#> 433 -0.5307146090
#> 434 -0.5207727150
#> 435 -0.5107596838
#> 436 -0.5006768830
#> 437 -0.4905256900
#> 438 -0.4803074914
#> 439 -0.4700236831
#> 440 -0.4596756697
#> 441 -0.4492648649
#> 442 -0.4387926907
#> 443 -0.4282605776
#> 444 -0.4176699643
#> 445 -0.4070222975
#> 446 -0.3963190317
#> 447 -0.3855616288
#> 448 -0.3747515584
#> 449 -0.3638902972
#> 450 -0.3529793286
#> 451 -0.3420201433